Update to @github/copilot-sdk 1.0.0-beta.10 and rename sandbox->environment, bash->virtualBash
- Bump @github/copilot-sdk from 0.3.0-preview.0 to 1.0.0-beta.10

- Adopt new SDK APIs: gitHubToken, mode: empty, createSessionFsProvider, getEvents(), ToolSet/BuiltInTools

- Rename sandbox -> environment (file rename, createSandbox -> createEnvironment, README prose)

- Rename bash -> virtualBash on session and createBash exports

- Remove obsolete 'filesystem isolation is incomplete' caveat (now fully contained by the runtime)
