+++ app/package.json
@@ -9,7 +9,7 @@
   "dependencies": {
     "@base-ui/react": "^1.3.0",
     "@fontsource-variable/geist": "^5.2.8",
-    "@github/copilot-sdk": "^0.3.0-preview.0",
+    "@github/copilot-sdk": "^1.0.0-beta.10",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "express": "^4.21.0",
